Fix readUserAnalaysis issue

+++ Sources/Mongli-Server/API/DreamAPI.swift
@@ -285,7 +285,7 @@ extension App {
             return next()
           }
 
-          guard let result = queryResult.first?["total"] as? Int32 else {
+          guard let result = queryResult.first?["total"] as? Int64 else {
             response.status(.internalServerError)
             return next()
           }
